scores for 7 variables

+++ metrics/sliced_wasserstein.py
@@ -149,9 +149,9 @@ def pyr_up(minibatch):  # matches cv2.pyrUp()
 
 def generate_laplacian_pyramid(minibatch, num_levels):
     if isinstance(minibatch, np.ndarray):
-        pyramid = [np.float32(minibatch)]
+        pyramid = [np.float16(minibatch)]
     else:
-        pyramid = [np.float32(minibatch.cpu())]
+        pyramid = [np.float16(minibatch.cpu())]
     for i in range(1, num_levels):
         pyramid.append(pyr_down(pyramid[-1]))
         pyramid[-2] -= pyr_up(pyramid[-1])
